+++ usvg/CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Coordinates resolving when units are `userSpaceOnUse`.
 - Groups regrouping.
 - Style attributes resolving on the root `svg` element.
+- `SmoothCurveTo` and `SmoothQuadratic` conversion.
 
 ### Removed
 - `Rect::transform`.